refactor(challenge): my-challenge/section DailyMission 3종을 daily-mission/ 하위폴더로 (구조정리 P4)
DailyMissionSection/InfoSection/SubmitSection → section/daily-mission/.
(RecommendedProgram 2·Doc 2는 3개 미만이라 과분리 지양하고 유지)
git rename + tsc-주도 import 갱신. 동작 변경 없음.
