+++ apps/web/src/domain/challenge/my-challenge/section/daily-mission/DailyMissionInfoSection.tsx
@@ -1,8 +1,8 @@
 import HybridLink from '@/common/HybridLink';
 import { UserChallengeMissionDetail } from '@/schema';
 import { BONUS_MISSION_TH } from '@/utils/constants';
-import OtVideo from '../../ot-mission/OtVideo';
-import ContentsDropdown from '../dropdown/ContentsDropdown';
+import OtVideo from '../../../ot-mission/OtVideo';
+import ContentsDropdown from '../../dropdown/ContentsDropdown';
 
 interface Props {
   missionDetail: UserChallengeMissionDetail;
