+++ apps/web/src/domain/challenge/my-challenge/section/daily-mission/DailyMissionSubmitSection.tsx
@@ -7,13 +7,13 @@ import { useOldCurrentChallenge } from '@/context/OldCurrentChallengeProvider';
 import { MyDailyMission, Schedule } from '@/schema';
 import { BONUS_MISSION_TH } from '@/utils/constants';
 import { useEffect, useState } from 'react';
-import BonusMissionInputSection from '../../bonus-mission/BonusMissionInputSection';
-import DailyMissionLinkInputSection from '../../daily-mission/DailyMissionLinkInputSection';
-import DailyMissionReviewSection from '../../daily-mission/DailyMissionReviewSection';
-import DailyMissionSubmitButton from '../../daily-mission/DailyMissionSubmitButton';
-import LinkChangeConfirmationModal from '../../modal/LinkChangeConfirmationModal';
-import OtMissionInputSection from '../../ot-mission/OtMissionInputSection';
-import LastMissionSubmitModal from './LastMissionSubmitModal';
+import BonusMissionInputSection from '../../../bonus-mission/BonusMissionInputSection';
+import DailyMissionLinkInputSection from '../../../daily-mission/DailyMissionLinkInputSection';
+import DailyMissionReviewSection from '../../../daily-mission/DailyMissionReviewSection';
+import DailyMissionSubmitButton from '../../../daily-mission/DailyMissionSubmitButton';
+import LinkChangeConfirmationModal from '../../../modal/LinkChangeConfirmationModal';
+import OtMissionInputSection from '../../../ot-mission/OtMissionInputSection';
+import LastMissionSubmitModal from '../../modal/LastMissionSubmitModal';
 
 interface Props {
   myDailyMission: MyDailyMission;
